refactor(proto): split alert raw config parsing
